pyproject.toml
[project]
name = "wikisearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bg_news_lg @ https://huggingface.co/sakelariev/bg_news_lg/resolve/main/bg_news_lg-3.5.4-py3-none-any.whl",
    "beautifulsoup4>=4.12.3",
    "docopt>=0.6.2",
    "faiss-cpu>=1.10.0",
    "fastapi>=0.115.8",
    "lmdb>=1.6.2",
    "lxml>=5.3.0",
    "mwparserfromhell>=0.6.6",
    "mysql-connector>=2.2.9",
    "nltk>=3.9.1",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.4.1",
    "spacy~=3.5.4",
    "tomli>=2.2.1",
    "types-tqdm>=4.67.0.20241221",
    "usearch>=2.16.9",
    "pyspellchecker>=0.8.2",
    "spellchecker>=0.4",
    "hunspell>=0.5.5",
    "dawg2>=0.13.1",
    "flask>=3.1.0",
    "uvicorn>=0.34.0",
    "elasticsearch>=8.17.1",
]

[tool.mypy]
disable_error_code = ["import-untyped"]

[tool.uv.sources]
bg-news-lg = { url = "https://huggingface.co/sakelariev/bg_news_lg/resolve/main/bg_news_lg-3.5.4-py3-none-any.whl" }
